adapter: Reject ws requests above a certain size
This commit adds logic to the websocket server to reject any request
above 2mb.

Fixes #17487
